Avoid empty slideshow thumbnail
If the first slide in a slideshow is text it won't have a thumbnail.
Look for a later slide that has a thumbnail and use that in this case.
This happens with Sprint when receiving an MMS with image+text, the
text is always the first slide in the slideshow.

Change-Id: Ib29acf07a533ff5531a1269792ff805fe8156b78